requirements.txt
email-validator
fastapi==0.38.1
requests
starlette==0.12.8
starlette-prometheus
tzlocal
uvicorn
structlog
python-json-logger
mock
mocker
pandas
XlsxWriter
mysql-connector-python
pytest
bcrypt
ConfigParser
xlrd
aiofiles